Dice Of Destiny - Roll To Riches (JavaScript Graphics)
October 3rd, 2025
I made this game called Dice Of Destiny back in December 2024, I didn't publish it since then, but here it is, one of my first ever longest lines of code, 1.9k Lines Of Code, almost 2,000, I'm happy how it turned out.

Game Concept:
Roll A Dice, you get a choice, to deposit your funds and then withdraw later, or keep it in the ATM. Another option you can do is the 50/50 option, you get a chance of earning x2 more income or lose it all, it's a gamble. Lastly, you can simply roll the dice and hope for the best, each number has it's very own perk, it can be good or bad.