Python Project (Rock, Paper, Scissors)
This file is my first ever project that I've done in my coding journey! Please remember to copy and paste the code onto any Python website, I personally recommend CodeHS, Python 3 on Sandbox, because it is more easier for this code to run on.

Note: The code may not be as good as expected, because this was my first ever project, so the chances of it being accurate is low, but please consider using this code to play the game, if you want to support me.